Resolving bugs in trendent and western digital
In Trendent I have added an alert check when we navigate to page.
In Western digital we already are waiting for 30 secs for
the page to reload and the radio buttons to turn on. Usually the change
happens within milliseconds. So I have just turned off the check on the
ssid text field.

BUG=chromium:330274, 330276
TEST=unittest

Change-Id: I8ae408ada68ff3d5244b42a9f45cd3d780e78447
Reviewed-on: https://chromium-review.googlesource.com/181732
Reviewed-by: Deepak Gopal <deepakg@chromium.org>
Commit-Queue: Deepak Gopal <deepakg@chromium.org>
Tested-by: Deepak Gopal <deepakg@chromium.org>