Restrict non-US APs to static configurators.
Currently all of non-US APs are static configurators.  This means
when creating an AP spec for a specialized channel, 48, if the
filter does not include the configurator type the dynamic APs will
be run a second time on the 5GHz band.  This doesn't provide any
extra data and greatly increases the time of the test.

TEST=Run the two control files.
BUG=chromium:331978

Change-Id: I15537b29c1a0404f6137cd772d8a30deb1abe4c5
Reviewed-on: https://chromium-review.googlesource.com/181850
Tested-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>