Touch FW linearity validator - Exception handling
There was a problem running the touch FW test on the robot where the
linearity validator would crash because sometimes the finger would
register a small tap before starting. As a result there
would be a brief contact before the actual gesture and the linearity
validator would crash due to there not being enough events in the first
touch.

This patch makes the linearity validator handle this failure more
gracefully instead of bailing on the whole test

BUG=chromium:332550
TEST=I ran the robot test on the offending system and the test worked
all the way through.  Previously it would never make it past a few
gestures before this problem cropped up.

Change-Id: I8a0e1a52a0e12b17c0840d27d4e914f6e3d04035
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/182054
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>