autotest: Refactoring and cleanup for LinuxRouter
In general, this patch tries to remove some of the invisible state and
opaque configuration built into the router.  As we've extended the
router to do more things, concepts like the "phytype" of the router no
longer make sense, because we can be multiple phytypes at once.
Fortunately, the phytype is easily inferred from the type of interface
we hope to configure.  We also had a number of interfaces to configure
various aspects of how the router worked that we never used and added
complexity.

Removed params dict for LinuxRouter/LinuxSystem.
Deprecated self.router field for LinuxRouter.
Removed LinuxRouter.create_wifi_device() since we now have
  hostapd/wpa_supplicant/ibss (phytype) specific method calls.
Removed LinuxRouter.destroy() method in favor of deconfig().

BUG=None
TEST=wifi_matfunc passes with these changes on a spring.

Change-Id: I1d77499b5e544813956e6f8018ca328c466046e8
Reviewed-on: https://chromium-review.googlesource.com/182636
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>