[Autotest] Avoid race conditions between special tasks and hqes.
Since the order of things that occur within a tick is:

-create an agent for a special task
-scheduler new jobs on ready hosts
-activate and run the special task, which changes the status of the host
removing it from the ready queue.

We sometimes find ourselves in a situation where a special task
scheduled through the frontend (eg: mass reverify) creates an agent
for a 'Ready' host, thereby making it unusable, but the host still gets
assigned to an hqe. This is because the host_scheduler till now, didn't
understand that an inactive special task could still invalidate a host
by assigning an agent to it, without actually changing it's status till later.
In such situations the hqe would sit around in a Queued state.

BUG=chromium:333122
TEST=Created special tasks for hosts through the frontend and checked that hosts
     don't get assigned to hqes.
DEPLOY=Scheduler

Change-Id: Icccb42e14ed86daeb2dcb38f3c492614fcd8d7ac
Reviewed-on: https://chromium-review.googlesource.com/182704
Tested-by: Prashanth B <beeps@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>