netprotos: Add initial unittests for Zeroconf module.
The ZeroconfDaemon class provides a partial python implementation of
the Zeroconf protocol used for integration testing of P2P features.
The implementation is based on the lansim simulator but it can be run
independently with a different underlying Host class with the same
interface.

This patch adds a new FakeHost and basic unit tests for the
ZeroconfDaemon class.

BUG=chromium:328078
TEST=./utils/unittest_suite.py --debug client.cros.netprotos.zeroconf_unittest

Change-Id: Ie081f8ba9a78bc42603d7397c4e9be5ebe439a2e
Reviewed-on: https://chromium-review.googlesource.com/182730
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>