netprotos: Process authoritative answers on mDNS queries
A mDNS query can contain a list of authoritative answers separated
from the list of known answers to the sent queries. This patch
makes the ZeroconfDaemon process those answers when they are
seen on a query packet and includes a unittest for that case.

BUG=chromium:328078
TEST=unittests and p2p_ServeFiles still passes.

Change-Id: I0609d35a2eaf08118d30f3ed68ed89754e803a42
Reviewed-on: https://chromium-review.googlesource.com/182731
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>