Moving cursor to the bottom-right corner
The original code doesn't work. Change to passing the active output
to the method to get the external framebuffer info and moving the cursor
to the bottom-right corner, which has fewer impact on the screen.

BUG=chromium:330804
TEST=manual
Run display_Resolution test mirrored mode on Link.
Checked that the cursor is in the bottom-right corner.

Change-Id: I608d012525ed6cf91f8fc5ce5d3bfe853a1430ab
Reviewed-on: https://chromium-review.googlesource.com/183433
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>