Touch FW test -- use shrunken device spec
For the gestures that require fingers to be potentially widely spread
apart, the robot needs to use the shrunken version of the device spec
(_min.p) so that we know the robot can actually span the distance.
These gestures are the one-finger-stationary gestures and the pinch
gestures.  This CL just changes their cmd line arguments to use the _min
spec instead of the regular full-screen spec in case it's too big.

BUG=chromium:336595
TEST=manually tested on a Link touchscreen and it worked great
CQ-DEPEND=Ibe5707aae3bbead727fd3491ba1d27985aa5b4c6

Change-Id: I4eb76857233575ab962730c6374558fa056cc7d1
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/183389
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>