Touch FW test: Add support for stationary/tap
The robot needs to know how to parse the gesture where it holds one
finger still and taps around it with a second finger.  This CL adds in
the parsing code and informs the FW test suite that the robot is now
capable of performing this gesture.

BUG=chromium:334460
TEST=Manually tested by running the FW test suite with the robot
CQ-DEPEND=I869c749aa35832bd4ca071d7db588bc903c48692

Change-Id: Ie6672a64ff8a5bdb51b5760e1b6bb88925db1c20
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/183531
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>