Autotest: Add new test to verify data integrity when using trim command
Add new hardware_TrimIntegrity test.

This test will write 1 GB of data and verify that trimmed data are gone and
untrimmed data are unaffected. The verification will be run in 5 passes with
0%, 25%, 50%, 75%, and 100% of data trimmed.

This test will raise an error only when untrimmed data are corrupt. It will
also report that trimmed data is zero'ed or not. If the target device does not
support trim, this test will log about that and exit without raising an error.

BUG=chromium:329736
TEST=Tested in link, falco, pit.
Test passed. Result show that trimmed data are zero'ed in falco, did not change
in link and ioctl does not support trim command in pit.

Change-Id: Ida23cc9aa05b639f1329236f0f6494c413fa55d0
Reviewed-on: https://chromium-review.googlesource.com/183712
Tested-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Puthikorn Voravootivat <puthik@chromium.org>