power_daily: Retrieve wifi attributes from global config
This pulls the wifi attributes from the shadow config instead of from the
machine attributes. Since there seems to be an allowance for the global_config
code to change and throw exceptions, this also moves the wired version of
power_LoadTest to before the wifi version. I remove the values for the wifi
parameters in the global_config so I can set the default value to None in the
case that a wifi network is not setup.

BUG=chromium:337028
TEST=run power_daily suite in the lab

Change-Id: I7148c5ff6919f91c95f31765d89c44e01bfd2c6a
Reviewed-on: https://chromium-review.googlesource.com/183731
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>