tcpdump_analyzer: add parsing of SSIDs in probe requests
We'd like to write some tests that verify that we scan for the
correct SSIDs. So give tcpdump_analyzer the ability to parse SSIDs
out of probe requests.

This CL also restructures the existing code a bit. The optional
matching group for MCS info is broken out of the main regexp.
The idea is that each of the optional fields can be parsed
independently, rather than constructing one giant regexp to
rule them all.

BUG=chromium:334721
TEST=network_WiFi_RateControl, manual

Manual testing
--------------
- Generated pcap containing broadcast probes, directed probes, and
  other traffic.
- Ran analyzer, printed out returned frames, and observed that the
  frames had the appropriate value for the probe_ssid field. (None
  for "other traffic", the empty string for broadcast probes, and
  the SSID for directed probes.)

Change-Id: I8f95c520f3fb27d4c5a693c4157435fc65fd8083
Reviewed-on: https://chromium-review.googlesource.com/183852
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>