autotest (wifi): add VisibleScan test
If we don't have any hidden SSIDs configured, then we should probe
for the broadcast SSID (and only the broadcast SSID).

This test starts a packet capture, connects to a visible network,
The test then verifies that a) the packet capture contains a 802.11
probe request for the broadcast SSID, and b) the packet capture does
not contain probe requests for any other SSIDs.

BUG=chromium:334721
TEST=manual

Manul testing
-------------
1. Run this test with a device in a shield box. Verify that the
   test passes.
2. Modify |pcap_filter| to filter out probe frames. Verify that
   the test fails.
3. Modify |pcap_filter| to select probe frames, but to allow
   frames from any MAC address. Open shield box. See that the
   test fails. (This depends on the environment; in my environment,
   there are plenty of probe requests for hidden SSIDs, from other
   clients.)

Change-Id: I930ee8eb91b6f73756ac8ef6bc5d5ef0ef5daaf8
Reviewed-on: https://chromium-review.googlesource.com/183875
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>