power_suspend: Make kernel WARNINGs end in TestWarn instead of TestFail
This patch changes all suspend/resume tests to throw a TestWarn instead
of a TestFail exception when detecting a kernel WARNING during the test
run. This will not close the tree if power_Resume is part of the BVT,
but will still auto-file bugs.

BUG=chromium:335799
TEST=Run power_Resume with a kernel that triggers WARNINGs, make sure
you get a "WARN" result. Ensure other kinds of test failures still
result in "FAIL".

Change-Id: I1645e038a58ea4799005e397c954afc2266874f0
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/183547
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>