autotest (wifi): add network_WiFi_HiddenScan test
If we have a hidden SSID configured, we should probe for both the hidden
SSID, and the broadcast SSID.

This tests starts a packet capture, and connects to a hidden network.
The test then verifies that a) the packet capture contains probe
request(s) for the hidden SSID and the broadcast SSID, and b) the
packet capture does not contain probes for other SSIDs.

While there: refactor some code, so that we can share it between
this test, and network_WiFi_VisibleScan.

BUG=chromium:338519
TEST=ran test, saw it pass

Change-Id: Icf8ca49e2806f035329aa7e7ca8a761f49f117c6
Reviewed-on: https://chromium-review.googlesource.com/184013
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>