graphics_GLMark2: Output FrameTime for all scenes
For each test, GLMark2 prints FPS and FrameTime to stdout.
A test is a scene plus its options.

Parse stdout for the test results and output the FrameTime value as a
perf_value.  The "description" field of the perf value is generated from
<scene>.<options>.  However, the "description" field has character
restrictions, so we massage the description using a python string
translator.

Also: Use "raw strings" for python regular expressions

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:330529
TEST=./run_remote_tests --remote=${ip} graphics_GLMark
  => Results now include per-scene FrameTime numbers

Change-Id: I581c0a8ded3eb12a2db7d41600a65db36d7c96dd
Reviewed-on: https://chromium-review.googlesource.com/183296
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>