platform_PartitionCheck : remove arch dependence when checking for mmcblk0
Rambi is an x86 system that uses mmcblk0. Change the logic in this test
to check for the existence of a nonremovable mmcblk0 and use that if found.

Ideally, we would be using a more robust method for determining what the
root device looks like, such as the json definition of the device.
See crbug.com/336654 for the installer.

Also cleaned up docstring params a bit and get rid of unused import stat.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chromium:338755
TEST=test_that <remote> platform_PartitionCheck
Run the test on Rambi (with mmc root device). Ensure it passes.
Run the test on Lumpy and Link as well, to check it doesn't break sda type
x86 devices.

Change-Id: Iad2e22c6a1167b2c36d8ae56db1959ece99756cb
Reviewed-on: https://chromium-review.googlesource.com/184108
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Commit-Queue: Benson Leung <bleung@chromium.org>