Move platform_Powerwash to "au" suite.
The platform_Powerwash test wipes the stateful partition removing
with it required programs like Python. After a platform_Powerwash
runs on a DUT, the lab will recover the device with a new
stateful partition. Nevertheless, when you run the test on your desk
as part of the BVT suite, you have to manually do that.

To avoid this problem while running the BVT suite, we move the test
to the AU suite.

BUG=chromium:338497
TEST=None

Change-Id: Ife4f44717e8c946b6f7161efcdec84a0a991ee89
Reviewed-on: https://chromium-review.googlesource.com/184437
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>