autoupdate_EndToEnd: Start private devserver with arbitrary port (take 2)
We switch to letting the private (Omaha mocking) devserver pick its own
port, instead of telling it which port to use. This depends on the
devserver's new feature allowing us to pass --port=0, instructing it to
allocate an available port, which is reported in its output log. This
complicates things slightly because, not knowing the port a priori, we
now need the devserver to tell us what it is via the --portfile.

Additional changes in this CL:

* We no longer attempt to kill any remote processes other than the one
  we started. This was a workaround meant to allow our port allocation
  scheme to keep working in the presence of stray devserver processes.

* There's no longer a point in using the port as indicator for whether
  a devserver is still alive: in some cases, we don't know even know
  what the port number is because the server did not report it.
  Instead, we use the remote pid as the sole test for the devserver's
  liveness.

* Devserver invocation is wrapped by timelimit(1), to ensure stray
  processes will not be hogging the host. To ensure we're killing the
  right process, we now use --pidfile to read the server's own PID
  (instead of that of timelimit).

* Renamed kill() into stop_devserver(), to emphasize the symmetry with
  start_devserver(); renamed a few other methods to better reflect their
  semantics and usage.

(This is a second attempt at merging this change that fixes a bug in the
previous attempt.)

BUG=chromium:333082
TEST=None

Change-Id: I9374ae5976e5749be2457d147369dcc41b770669
Reviewed-on: https://chromium-review.googlesource.com/187467
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Tested-by: Gilad Arnold <garnold@chromium.org>