make platform_CryptohomeMount experimental
Per discussion on the bug, retries for this test have not fixed the test
flake because the failures are due to a dead cryptohomed which doesn't
restart fast enough for the retried test to succeed.

Since this bug is already under investigation, and is a leading cause of
CQ failure at the moment, marking it as experimental.

BUG=chromium:303677
TEST=None

Change-Id: I37546e4be53a712b06956f5118fa8ffdcc42c469
Reviewed-on: https://chromium-review.googlesource.com/184557
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Reviewed-by: Will Drewry <wad@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>