Update the robot wrapper for new click format
The click.py argument format is changing to accomodate a finger size
parameter at the end.  This means that for physical clicks the robot
needs to add a integer on to the end.  This patch changes the robot
wrapper for the Touch firmware test to follow this new format for
physical clicks.

BUG=chromium:332473
TEST=Ran the FW test physical clicks with the new patches
CQ-DEPEND=Idfa4c7a7f783d3aa28c65d061a378f533d2ae4c1

Change-Id: I623e3f5d9273cf24c3af712db63dd03eb13139a2
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/184704
Reviewed-by: Dennis Kempin <denniskempin@chromium.org>