autotest: Wait for connection and frequency changes during roaming
In the past, during roaming tests we waited for WiFi services to reach a
connected state, then asserted that they were connected on a particular
frequency.  If this assertion failed, then we would call it a roaming
failure.  This breaks down when a service fails to transition to
a disconnected state fast enough, for instance if the DEAUTH
notification is lost, or handled internally by the driver during roaming
situations.  Fix this by waiting for a reasonable period of time
for both the service to enter an online state, and the frequency to
change.

TEST=network_WiFi_Roam passes reliably (20 times without failure) with
this change. wifi_matfunc continues to pass with this change.
BUG=chromium:340836

Change-Id: Ib390ed1ccad5954b286f582c35f850d87af9848b
Reviewed-on: https://chromium-review.googlesource.com/185050
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>
Reviewed-by: <dnkim1205@gmail.com>