autotest Allow running chaos tests without AFE access
Previously, chaos test could be run against devices(DUTs, APs and packet
captures) that are local to developers environment by providing
the host names of these devices in the control file. This support was broken
with the introduction of the patch "Filter APs based on location relative to
the lab". It assumes AFE is accessible, but not all APs are in AFE. So fix this
by skipping calls to AFE if hostnames of APs are provided.

BUG=None
TEST=network_WiFi_ChaosConnectDisconnect/control.debug$ passes against an AP
not accessible via AFE

Change-Id: I13f4b762c9e0fa2d6dcd687ca58cca2659f94ff0
Signed-off-by: Ashok Nagarajan <asnagarajan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/185284