pseudomodem: detect pseudomodem subprocess crashes.
pseudomodem_context launches a subprocess to run pseudomodem. This CL adds a
signal handler to the parent process to detect the scenario when the
pseudomodem subprocess dies earlier than expected.

Before this CL, the main process would ignore the status of the child process.
So, a test using the pseudomodem would not know that pseudomodem had crashed.

While there:
  - When run standalone, make the logging from the main process prettier.

BUG=chromium:337871
TEST=(1) Run pseudomodem standalone (run_pseudomodem.py). kill the subprocess
         launched, and make sure that run_pseudomodem.py exits.
     (2) Run network_3GModemControl.pseudomodem

Change-Id: I5792984ef8901febb98672c2ff48555328f4ef9f
Reviewed-on: https://chromium-review.googlesource.com/184967
Reviewed-by: Ben Chan <benchan@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>