[autotest] Add some code useful to monitoring/deploying to site_utils/lib
I'm not quite sure where to put this code, so I'm just creating a home
for it?

infra.py currently contains code to run a command on one of our servers
in the optimal way depending on the host that you're running on.  It
also contains functions to pull the list of all servers used for certain
purposes (scheduler, database, etc.) in the lab.

BUG=chromium:340454
TEST=ran functions, verified results

Change-Id: I9e65f6f7a6ebcb6b9f1734670a87058f44061c33
Reviewed-on: https://chromium-review.googlesource.com/184978
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>
Tested-by: Alex Miller <milleral@chromium.org>