pseudomodem: Surface pseudomodem process exceptions in autotest logs.
pseudomodem runs in a separate process and redirects its logs to syslog. When a
test is run, if pseudomodem crashes, it is desirable to surface the reason to
the autotest logs. This will help our tools to dedup bugs better.

With this CL, the pseudomodem child process communicates its error status to the
launching |pseudomodem_context| which then raises an exception including the
error status.

BUG=chromium:337869
TEST=Introduce an error in pseudomodem by raising an exception in |Modem|
     constructor. Then
     (1) Run pseudomodem standalone and check the exception raised.
     (2) Run a pseudomodem test and check the error reason on stdout.
     (3) Ensure no files are left behind matching /tmp/pseudomodem_*

Change-Id: Iaa3483292968c1cb03680c9908ec5c102a03c356
Reviewed-on: https://chromium-review.googlesource.com/185082
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>