Add support for profiling using minimal functionality from Telemetry.
Telemetry benchmarks are quite flaky at this point in ChromeOS. So we
are adding a new mode to this test to capture the Chrome profile using
the minimal support from Telemetry. For that we need to control Chrome
directly which needs to be done with a client test. The client test
will ask Chrome to display the benchmarks pages directly instead of
using Telemetry "page sets" and "measurements" support.

BUG=None
TEST=Tested through test_that using client test (work in progress).
     Verified the profile is captured and only one instance of the
     profiler is executed.

Change-Id: I4140aa41b6930281fff1b1013c9664b5747a9516
Reviewed-on: https://chromium-review.googlesource.com/185123
Reviewed-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>