Fix list_phys() to return an empty list
If "iw list" returns an nothing (if no wireless interfaces are
reported), list_phys() should return an empty list.  This CL
fixes the current implementation so this is the case.

BUG=chromium:341892
TEST=Do "rmmod <wifi_module>" on a DUT, then run a WiFi test.
Observe that the test fails with "No interfaces managed by shill"
and attempts a reboot of the DUT.

Change-Id: I4ab2ef6fd50b93b29b011576f834e38e9bea4cff
Reviewed-on: https://chromium-review.googlesource.com/185342
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>