Touch FW test: add robotic fat-finger support
The fingertip size selection process for the touchbot wrapper for the
touch FW test was somewhat crude before.  All the gestures that it
supported all used the same size fingertips or no fingertips so it was
good enough to just have pretty basic logic.

This CL revamps the fingertip size selection logic to allow for
fat-finger tests, and enables the two-fat-finger-tracking test.  Most of
the logic for this gesture already existed in the regular
two-finger-tracking test, so the majority of the CL is just setting up
the switching code to make it work.  It's a little overkill for this
test, but should enable multi-sized gestures when we add them in the
future, which is coming soon.

BUG=chromium:334460
TEST=manually ran the FW test

Change-Id: I99f45f36d493d865b72b9e5499227e82a1e7a769
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/185391