Remove unnecessary sync calls from setup phase again.
This test runs 'sync' during the setup phase of the build. This causes the
build to pause and wait for all pending writes to complete. This step delays
the build unnecessarily. In my tests I saw this step causing the build to hang
for several minutes because other processes were writing to the disk.

This CL was previously committed in CL:2188 but was accidentally partially
reverted during CL:4664.

BUG=chromium-os:16218
TEST=pre-cq-group trybot run.

Change-Id: I7b6af9fff303caad3e9a46d9c5f0869c8c81ef0c
Reviewed-on: https://chromium-review.googlesource.com/185410
Tested-by: David James <davidjames@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>