login_UserPolicyKeys and logging_UserCrash are now more hermetic
login_UserPolicyKeys was leaving the session_manager believing
that a user session was still underway at the end of the test.
logging_UserCrash was assuming that the device was at the sign-in
screen without actually ensuring that this was so. Both were
wrong.

BUG=chromium:342195
TEST=run login_UserPolicyKeys and logging_UserCrash in that order
STATUS=Fixed

Change-Id: I1eca08b8a367ddd2ca7483040ed15d05ac541079
Reviewed-on: https://chromium-review.googlesource.com/185640
Commit-Queue: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>