platform_CryptohomeMount: drop experimental
The test was flaking on snow due to improper
timeout exception handling on DBus calls exceeding
a 10s wait.  Originally, coredumping was added to
diagnose overly long calls -- those were due to
hardware contention (TPM).  The timeout increase
was meant to accomodate, but the improper handling
of a nested timeout resulted in continued flake.

With that fixed and the good track record on snow
since, I'm moving this back out of experimental.

BUG=chromium:303677
TEST=n/a

Change-Id: If3c10661c2a8f9d00025161de51b612d25d9598b
Reviewed-on: https://chromium-review.googlesource.com/185585
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: Will Drewry <wad@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>