Navigate a pre-defined set of URLs for Chrome profile collection.
The purpose of this test is to exercise chrome with a meaningful set
of pages while a profile of Chrome is captured. It also aims at using
the minimum set of functionality from Telemetry since Telemetry is not
very stable on ChromeOS at this point.

This test is designed to be called from the telemetry_AFDOGenerate
server test. The server test will start the "perf" profiling tool on
the DUT before starting this test. It will also capture the chrome
profile and upload it to Google Storage to be used for an optimized
build of Chrome.

BUG=None
TEST=Tested with test_that and in combination with telementry_AFDOGenerated
     server test.

Change-Id: Ibe94cef3000289f93b01fb82c56b8bb00f7ba802
Reviewed-on: https://chromium-review.googlesource.com/185735
Reviewed-by: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>