Increase the wrong pixel margin and use it only on mirrored mode
Increase the wrong pixel margin to 100 because in some case the cursor
may change to a zoom icon, which is much larger. And change to use this
margin only on mirrored mode.

BUG=chromium:336731
TEST=manual
Run display_Resolution test on Link. Passed on extended and mirrored modes.

Change-Id: I6d75a618acf2f8167038dd53f2b6d90922704dc8
Reviewed-on: https://chromium-review.googlesource.com/185827
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>