firmware_TouchMTB: create the version info and scp it to test machine
It is helpful to add version information in the test results so that
we know the version of the test suite that generates the results.

In this patch, a script is provided to generate the version
information and scp it to the test machine. In next patch, there
should be some automatic way to invoke the script. For example,
add a hook in test_that script which will replace run_remote_tests.sh

BUG=chromium:315787
TEST=Perform the steps in chroot
(cr) $ cd .../site_tests/firmware_TouchMTB
(cr) $ ./test_that_hook.sh -r ${CHROMEBOOK_IP_ADDRESS}
As an example:
(cr) $ ./test_that_hook.sh -r 100.20.300.123
On the chromebook machine,
$ cat /usr/local/autotest/tests/firmware_TouchMTB/.version
It should print something like
version: 2014-01-23.16:18:24 1feddb865cf3f9c158d8cca17a6e68c71b500571

Change-Id: I652c0aa7e7c3ec6f42904cdfc21a8905896fc9ab
Reviewed-on: https://chromium-review.googlesource.com/185816
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>