Set Touch FW test to use visible clicks in mtplot
The ability to make visible marks for clicks has just been added to
mtplot.  This allows the FW test to record a better version of the
mtplot logs where you can actually see everything that happens.

Previously clicks clear the display, which can sometimes mean we lose a
lot of debugging info.  With this change, clicks don't clear the screen,
but are shown as rectangles in the output image.

BUG=chromium:342433
TEST=I manually ran the FW test suite and clicked the pad during a test,
the image showed the little boxes instead of clearing the display.
CQ-DEPEND=Ic67775cc9dbbf73858d1d7c9051ea6d8145bc6c4

Change-Id: I1d6bed87188f8dc84921ea989ad2db5793f0596e
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/186107