pseudomodem: Make a modem manager process running a fatal error.
PseudoModemManagerContext used to warn the user if a process for
cromo / ModemManager was still running after the service is stopped. This CL
makes this error fatal, as it causes a race between the different modem managers
and pseudomodem. Also, we poll for some time for the processes to die, since
ModemManager can take a while to cleanly shutdown after the service is stopped.

BUG=chromium:341943
TEST=(1) Run network_3GModemPresent.pseudomodem successfully.
     (2) On the DUT,
           stop cromo
           stop modemmanager
           sudo -u modem ModemManager --debug
         Then run network_3GModemPresent.pseudomodem. It should fail after a
         while with an error that says that ModemManager could not be shut down.

Change-Id: I3cc4772404929c27780ee8cd2c2c9446702a8ecc
Reviewed-on: https://chromium-review.googlesource.com/186151
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>