Center the robot after changing the fingertips
The gestures immediately home to their first position.  This is fine as
long as there's nothing in the way, as is the case when the robot's hand
never gets too far away from the pad.  However, if the robot just went
to the fingertip nest to get a new set of fingertips, it's very possible
that it might clip the edge of the nest if the wrong gesture comes next,
the device is too close to the nest, or something is just in the way.

To try and alleviate this risk as much as possible, this CL adds a call
to the reset.py gesture after every time the robot gets a new set of
fingertips.  This gesture causes the robot arm to lift up to a safe
height, move over the center of the touch device and then lower down.
As a result, this will avoid clipping the nest if it's too close, etc.

Additionally, this shouldn't add much time to the run since the arm
needs to move over there one way or the other, this just makes it lift
up first.

BUG=chromium:334073
TEST=ran the FW test and watched it work

Change-Id: I0f7114555439908c2d44136b61e0d1cf111c2ded
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/186446