autotest (dhcp, vpn): remove flaky tests from bvt
The DHCP and VPN tests have been flaky recently, and it's going
to take some time to get to the bottom of those problems. Until
then, take them out of the BVT, so we don't hold up the CQ.

BUG=chromium:343299
TEST=manual

Manual test: ran
  ~/trunk/src/third_party/autotest/files/site_utils/suite_enumerator.py \
    -a ~/trunk/src/third_party/autotest/files bvt
checked that the list did not include any dhcp or vpn tests.

Change-Id: I635ae43f224df7dc1526420b7cc9710df6339cd1
Reviewed-on: https://chromium-review.googlesource.com/186183
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>