Add kernel_CrosECSysfs to the kernel_daily_regression suite
This test verifies that userspace can talk to the Chrome OS EC (on systems
that actually have one). We want to keep this working, so make sure to test
it fairly often.

BTW, marking it as experimental until I'm certain it's working as intended
everywhere.

BUG=chromium:334014
TEST=manual

I ran the suite locally using test_that. With this CL it included
kernel_CrosECSysfs, which passed. A bunch of other stuff failed (both with
and without this CL), but nothing new.

Change-Id: I7519f10fa4906134069e1a0a2568edfa0297bdf3
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/189241
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>