chaos suspend resume real world test
This patch would add a new test case to the chaos test, where you would check
the connection to the AP with suspend resume cycle involved. The DUT will ping
the work client before and after the suspend-resume to check the liveliness of
the connection to the AP. This test case will also capture the failure of DUT
to connect to the AP within 15 seconds.

While there, do some code cleanups with ConnectionDuration class:
    1. switch to public interface for getting WiFiClient.wifi_ip
    2. do not break lines on logging

TEST=ran control.suspend and check if DUT is pinging the packet capturer
sucessfully before and after suspend
BUG=chromium:308215

Change-Id: I715a7784fb1f6ef19df1b8e0653e35a4289229f7
Signed-off-by: Ashok Nagarajan <asnagarajan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/185078
Reviewed-by: Christopher Wiley <wiley@chromium.org>